Fix screen props for all missing screens
