+++ src/screens/doorToDoor/tunnel/doorSelection/TunnelDoorSelectionScreen.tsx
@@ -2,14 +2,17 @@ import React, { FunctionComponent, useEffect, useState } from 'react'
 import { Image, StyleSheet, View, Text, Alert } from 'react-native'
 import SafeAreaView from 'react-native-safe-area-view'
 import DoorToDoorRepository from '../../../../data/DoorToDoorRepository'
-import { DoorToDoorTunnelStartScreenProp, Screen } from '../../../../navigation'
+import { Screen } from '../../../../navigation'
+import { DoorToDoorTunnelModalNavigatorScreenProps } from '../../../../navigation/DoorToDoorTunnelModalNavigator'
 import { Colors, Spacing, Typography } from '../../../../styles'
 import i18n from '../../../../utils/i18n'
 import { PrimaryButton, SecondaryButton } from '../../../shared/Buttons'
 import LoadingOverlay from '../../../shared/LoadingOverlay'
 import { TouchablePlatform } from '../../../shared/TouchablePlatform'
 
-const TunnelDoorSelectionScreen: FunctionComponent<DoorToDoorTunnelStartScreenProp> = ({
+type TunnelDoorSelectionScreenProps = DoorToDoorTunnelModalNavigatorScreenProps<'TunnelDoorSelectionScreen'>
+
+const TunnelDoorSelectionScreen: FunctionComponent<TunnelDoorSelectionScreenProps> = ({
   navigation,
   route,
 }) => {
