+++ src/screens/personalInformation/notifications/NotificationMenuScreen.tsx
@@ -1,11 +1,14 @@
 import React from 'react'
 import { StyleSheet, Text } from 'react-native'
 import SafeAreaView from 'react-native-safe-area-view'
-import { NotificationMenuScreenProps, Screen } from '../../../navigation'
+import { Screen } from '../../../navigation'
+import { ProfileModalNavigatorScreenProps } from '../../../navigation/ProfileModalNavigator'
 import { Colors, Spacing, Typography } from '../../../styles'
 import i18n from '../../../utils/i18n'
 import ProfileSettingsItem from '../../profile/ProfileSettingsItem'
 import CircularIcon from '../../shared/CircularIcon'
+
+type NotificationMenuScreenProps = ProfileModalNavigatorScreenProps<'NotificationMenu'>
 
 const NotificationMenuScreen = (props: NotificationMenuScreenProps) => {
   const onLocal = () => {
