+++ src/screens/personalInformation/notifications/NotificationsScreen.tsx
@@ -13,7 +13,7 @@ import {
   GetNotificationsInteractorResult,
 } from '../../../core/interactor/GetNotificationsInteractor'
 import PersonalInformationRepository from '../../../data/PersonalInformationRepository'
-import { NotificationsScreenProps } from '../../../navigation'
+import { ProfileModalNavigatorScreenProps } from '../../../navigation/ProfileModalNavigator'
 import { Colors, Spacing, Typography } from '../../../styles'
 import i18n from '../../../utils/i18n'
 import { AlertUtils } from '../../shared/AlertUtils'
@@ -101,6 +101,8 @@ const NotificationsContent = (
     </>
   )
 }
+
+type NotificationsScreenProps = ProfileModalNavigatorScreenProps<'Notifications'>
 
 const NotificationsScreen = (props: NotificationsScreenProps) => {
   const [statefulState, setStatefulState] = useState<
