+++ src/screens/phonePollDetail/PhonePollDetailScreen.tsx
@@ -9,7 +9,6 @@ import { View, StyleSheet } from 'react-native'
 import { StatefulView, ViewState } from '../shared/StatefulView'
 import { CloseButton } from '../shared/NavigationHeaderButton'
 import ModalOverlay from '../shared/ModalOverlay'
-import { PhonePollDetailScreenProps } from '../../navigation'
 import PhoningCampaignRepository from '../../data/PhoningCampaignRepository'
 import PhonePollDetailScreenLoaded from './PhonePollDetailScreenLoaded'
 import PhonePollDetailInterruptionModalContent from './PhonePollDetailInterruptionModalContent'
@@ -22,6 +21,9 @@ import {
 } from '../../core/interactor/GetPhonePollDetailResourcesInteractor'
 import { AlertUtils } from '../shared/AlertUtils'
 import { ViewStateUtils } from '../shared/ViewStateUtils'
+import { PhoningSessionModalNavigatorScreenProps } from '../../navigation/PhoningSessionModalNavigator'
+
+type PhonePollDetailScreenProps = PhoningSessionModalNavigatorScreenProps<'PhonePollDetail'>
 
 const PhonePollDetailScreen: FunctionComponent<PhonePollDetailScreenProps> = ({
   route,
@@ -116,8 +118,6 @@ const PhonePollDetailScreen: FunctionComponent<PhonePollDetailScreenProps> = ({
               satisfactionQuestions={
                 resources.configuration.satisfactionQuestions
               }
-              route={route}
-              navigation={navigation}
             />
           </>
         )}
