+++ src/screens/phoningSessionLoader/PhoningSessionLoaderScreen.tsx
@@ -8,7 +8,8 @@ import {
   PhoningSessionNoNumberError,
 } from '../../core/errors'
 import PhoningCampaignRepository from '../../data/PhoningCampaignRepository'
-import { PhoningSessionLoaderScreenProps, Screen } from '../../navigation'
+import { Screen } from '../../navigation'
+import { PhoningSessionModalNavigatorScreenProps } from '../../navigation/PhoningSessionModalNavigator'
 import { Colors, Spacing, Typography } from '../../styles'
 import i18n from '../../utils/i18n'
 import { CloseButton } from '../shared/NavigationHeaderButton'
@@ -16,6 +17,8 @@ import { VerticalSpacer } from '../shared/Spacer'
 import { StatefulView, ViewState } from '../shared/StatefulView'
 import { usePreventGoingBack } from '../shared/usePreventGoingBack.hook'
 import { ViewStateUtils } from '../shared/ViewStateUtils'
+
+type PhoningSessionLoaderScreenProps = PhoningSessionModalNavigatorScreenProps<'PhoningSessionLoader'>
 
 const PhoningSessionLoaderScreen: FunctionComponent<PhoningSessionLoaderScreenProps> = ({
   navigation,
