+++ src/screens/phoningSessionLoaderPermanentCampaign/PhoningSessionLoaderPermanentCampaignScreen.tsx
@@ -8,10 +8,7 @@ import { Text, StyleSheet } from 'react-native'
 import SafeAreaView from 'react-native-safe-area-view'
 import { PhoningSession } from '../../core/entities/PhoningSession'
 import PhoningCampaignRepository from '../../data/PhoningCampaignRepository'
-import {
-  PhoningSessionLoaderPermanentCampaignScreenProp,
-  Screen,
-} from '../../navigation'
+import { Screen } from '../../navigation'
 import { Colors, Spacing, Typography } from '../../styles'
 import i18n from '../../utils/i18n'
 import { PrimaryButton } from '../shared/Buttons'
@@ -20,8 +17,11 @@ import { FlexibleVerticalSpacer, VerticalSpacer } from '../shared/Spacer'
 import { PhoningSessionNavigationData } from '../shared/PhoningSessionNavigationData'
 import LoadingOverlay from '../shared/LoadingOverlay'
 import { AlertUtils } from '../shared/AlertUtils'
+import { PhoningSessionModalNavigatorScreenProps } from '../../navigation/PhoningSessionModalNavigator'
 
-const PhoningSessionLoaderPermanentCampaignScreen: FunctionComponent<PhoningSessionLoaderPermanentCampaignScreenProp> = ({
+type PhoningSessionLoaderPermanentCampaignScreenProps = PhoningSessionModalNavigatorScreenProps<'PhoningSessionLoaderPermanentCampaign'>
+
+const PhoningSessionLoaderPermanentCampaignScreen: FunctionComponent<PhoningSessionLoaderPermanentCampaignScreenProps> = ({
   navigation,
   route,
 }) => {
