+++ src/screens/phoningSessionNoNumberAvailable/PhoningSessionNoNumberAvailableScreen.tsx
@@ -1,13 +1,15 @@
 import React, { FunctionComponent } from 'react'
 import { Text, StyleSheet, View } from 'react-native'
 import SafeAreaView from 'react-native-safe-area-view'
-import { PhoningSessionNoNumberAvailableScreenProps } from '../../navigation'
 import { Colors, Spacing, Typography } from '../../styles'
 import { usePreventGoingBack } from '../shared/usePreventGoingBack.hook'
 import { TertiaryButton } from '../shared/Buttons'
 import { VerticalSpacer } from '../shared/Spacer'
 import i18n from '../../utils/i18n'
 import CircularIcon from '../shared/CircularIcon'
+import { PhoningSessionModalNavigatorScreenProps } from '../../navigation/PhoningSessionModalNavigator'
+
+type PhoningSessionNoNumberAvailableScreenProps = PhoningSessionModalNavigatorScreenProps<'PhoningSessionNoNumberAvailable'>
 
 const PhoningSessionNoNumberAvailableScreen: FunctionComponent<PhoningSessionNoNumberAvailableScreenProps> = ({
   navigation,
