+++ src/screens/profile/ProfileLoginScreen.tsx
@@ -1,7 +1,10 @@
 import React, { FC } from 'react'
 
-import { Screen, ProfileLoginScreenProps } from '../../navigation'
+import { Screen } from '../../navigation'
+import { ProfileModalNavigatorScreenProps } from '../../navigation/ProfileModalNavigator'
 import LoginScreen from '../authentication/LoginScreen'
+
+type ProfileLoginScreenProps = ProfileModalNavigatorScreenProps<'ProfileLogin'>
 
 const ProfilLoginScreen: FC<ProfileLoginScreenProps> = ({ navigation }) => {
   return (
